Moved AssertQuery to TestBase.
